Remove the need for job.wait() and job.Stop() to lock the store explicitly.
Initially, the lock needs to be acquired by each method in order to avoid race conditions when the job is stopped.
However, Go returns an exit code of -1 to indicate that a process is stopped by a signal. This can be used to update the job values in the store
without acquiring a lock explicitly